# python-flask-server-generated/gen/api/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing beverages, labels, ingredients, container types, and
    unit measures in a beverage e-commerce platform.
  title: Beverage E-commerce API
  version: 1.0.0
servers:
- url: https://api.beverage-ecommerce.com/v2
security:
- BearerAuth:
  - read:beverages
  - write:beverages
  - read:brands
  - write:brands
  - read:labels
  - write:labels
  - read:ingredients
  - write:ingredients
  - read:container-types
  - write:container-types
...
paths:
...
  /beverages/{id}/labels:
    get:
      description: Retrieves a list of all labels.
      operationId: get_labels
      parameters:
      - description: Unique identifier for the beverages.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Label'
                type: array
                x-content-type: application/json
          description: A list of labels.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - read:labels
      summary: Retrieve all labels
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-accepts: application/json
    post:
      description: Creates a new label.
      operationId: create_label
      parameters:
      - description: Unique identifier for the beverages.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/id_labels_body'
        required: true
      responses:
        "201":
          description: Label created successfully.
          headers:
            Location:
              description: URL to the newly created label.
              explode: false
              schema:
                format: uri
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:labels
        - read:labels
      summary: Create a new label
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-contentType: application/json
      x-accepts: application/json
  /beverages/{beverageId}/labels/{id}:
    delete:
      description: Deletes a label.
      operationId: delete_label
      parameters:
      - description: Unique identifier for the beverage.
        explode: false
        in: path
        name: beverageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the label.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Label deleted successfully.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:labels
        - read:labels
      summary: Delete a label
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-accepts: application/json
    get:
      description: Retrieves details of a specific label.
      operationId: get_label_by_id
      parameters:
      - description: Unique identifier for the beverage.
        explode: false
        in: path
        name: beverageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the label.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
          description: Label details.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - read:labels
      summary: Retrieve a label by ID and beverage ID
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-accepts: application/json
    patch:
      description: Update existing label.
      operationId: update_label
      parameters:
      - description: Unique identifier for the beverage.
        explode: false
        in: path
        name: beverageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the label.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/labels_id_body_1'
        required: true
      responses:
        "200":
          description: Label updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:labels
        - read:labels
      summary: Partially update a label
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-contentType: application/json
      x-accepts: application/json
    put:
      description: Create or Replace an existing label.
      operationId: replace_label
      parameters:
      - description: Unique identifier for the beverage.
        explode: false
        in: path
        name: beverageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the label.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/labels_id_body'
        required: true
      responses:
        "200":
          description: Label updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:labels
        - read:labels
      summary: Update a label
      tags:
      - label
      x-openapi-router-controller: swagger_server.controllers.label_controller
      x-contentType: application/json
      x-accepts: application/json
...
  /ingredients:
    get:
      description: Retrieves a list of all ingredients.
      operationId: get_ingredients
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Ingredient'
                type: array
                x-content-type: application/json
          description: A list of ingredients.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
      security:
      - BearerAuth:
        - read:ingredients
      summary: Retrieve all ingredients
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-accepts: application/json
    post:
      description: Creates a new ingredient.
      operationId: create_ingredient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ingredients_body'
        required: true
      responses:
        "201":
          description: Ingredient created successfully.
          headers:
            Location:
              description: URL to the newly created ingredient.
              explode: false
              schema:
                format: uri
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
      security:
      - BearerAuth:
        - write:ingredients
        - read:ingredients
      summary: Create a new ingredient
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-contentType: application/json
      x-accepts: application/json
  /ingredients/{id}:
    delete:
      description: Deletes an ingredient.
      operationId: delete_ingredient
      parameters:
      - description: Unique identifier for the ingredient.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Ingredient deleted successfully.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:ingredients
        - read:ingredients
      summary: Delete an ingredient
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-accepts: application/json
    get:
      description: Retrieves details of a specific ingredient.
      operationId: get_ingredient_by_id
      parameters:
      - description: Unique identifier for the ingredient.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
          description: Ingredient details.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - read:ingredients
      summary: Retrieve an ingredient by ID
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-accepts: application/json
    patch:
      description: Updates an existing ingredient.
      operationId: update_ingredient
      parameters:
      - description: Unique identifier for the ingredient.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ingredients_id_body_1'
        required: true
      responses:
        "200":
          description: Ingredient updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:ingredients
        - read:ingredients
      summary: Update an ingredient
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-contentType: application/json
      x-accepts: application/json
    put:
      description: Create or Replace an existing ingredient.
      operationId: replace_ingredient
      parameters:
      - description: Unique identifier for the ingredient.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ingredients_id_body'
        required: true
      responses:
        "200":
          description: Ingredient updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:ingredients
        - read:ingredients
      summary: Update an ingredient
      tags:
      - ingredient
      x-openapi-router-controller: swagger_server.controllers.ingredient_controller
      x-contentType: application/json
      x-accepts: application/json
...
components:
...
  schemas:
...
    Label:
      example:
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          description: Unique identifier for the label.
          format: uuid
          type: string
        name:
          description: Name of the label.
          type: string
        description:
          description: Detailed description of the label.
          type: string
      required:
      - id
      - name
      type: object
    Ingredient:
      example:
        name: name
        allergenInfo: allergenInfo
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          description: Unique identifier for the ingredient.
          format: uuid
          type: string
        name:
          description: Name of the ingredient.
          type: string
        allergenInfo:
          description: Allergen information.
          type: string
      required:
      - id
      - name
      type: object
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponse_error'
      type: object
...
    id_labels_body:
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - name
      type: object
    labels_id_body:
      properties:
        name:
          type: string
        description:
          type: string
      type: object
    labels_id_body_1:
      properties:
        name:
          type: string
        description:
          type: string
      type: object
...
    ingredients_body:
      properties:
        name:
          type: string
        allergen_info:
          type: string
      required:
      - name
      type: object
    ingredients_id_body:
      properties:
        name:
          type: string
        allergen_info:
          type: string
      type: object
    ingredients_id_body_1:
      properties:
        name:
          type: string
        allergen_info:
          type: string
      type: object
...
    ErrorResponse_error:
      properties:
        code:
          description: HTTP status code.
          type: integer
        message:
          description: Error message.
          type: string
      required:
      - code
      - message
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth