# python-flask-server-generated/swagger_server/controllers/label_controller.py
import connexion
import six

from swagger_server.models.error_response import ErrorResponse  # noqa: E501
from swagger_server.models.id_labels_body import IdLabelsBody  # noqa: E501
from swagger_server.models.label import Label  # noqa: E501
from swagger_server.models.labels_id_body import LabelsIdBody  # noqa: E501
from swagger_server.models.labels_id_body1 import LabelsIdBody1  # noqa: E501
from swagger_server import util


def create_label(body, id):  # noqa: E501
    """Create a new label

    Creates a new label. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param id: Unique identifier for the beverages.
    :type id: 

    :rtype: None
    """
    if connexion.request.is_json:
        body = IdLabelsBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def delete_label(beverage_id, id):  # noqa: E501
    """Delete a label

    Deletes a label. # noqa: E501

    :param beverage_id: Unique identifier for the beverage.
    :type beverage_id: 
    :param id: Unique identifier for the label.
    :type id: 

    :rtype: None
    """
    return 'do some magic!'


def get_label_by_id(beverage_id, id):  # noqa: E501
    """Retrieve a label by ID and beverage ID

    Retrieves details of a specific label. # noqa: E501

    :param beverage_id: Unique identifier for the beverage.
    :type beverage_id: 
    :param id: Unique identifier for the label.
    :type id: 

    :rtype: Label
    """
    return 'do some magic!'


def get_labels(id):  # noqa: E501
    """Retrieve all labels

    Retrieves a list of all labels. # noqa: E501

    :param id: Unique identifier for the beverages.
    :type id: 

    :rtype: List[Label]
    """
    return 'do some magic!'


def replace_label(body, beverage_id, id):  # noqa: E501
    """Update a label

    Create or Replace an existing label. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param beverage_id: Unique identifier for the beverage.
    :type beverage_id: 
    :param id: Unique identifier for the label.
    :type id: 

    :rtype: None
    """
    if connexion.request.is_json:
        body = LabelsIdBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def update_label(body, beverage_id, id):  # noqa: E501
    """Partially update a label

    Update existing label. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param beverage_id: Unique identifier for the beverage.
    :type beverage_id: 
    :param id: Unique identifier for the label.
    :type id: 

    :rtype: None
    """
    if connexion.request.is_json:
        body = LabelsIdBody1.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'