# python-flask-server-generated/swagger_server/models/ingredient.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Ingredient(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, allergen_info: str=None):  # noqa: E501
        """Ingredient - a model defined in Swagger

        :param id: The id of this Ingredient.  # noqa: E501
        :type id: str
        :param name: The name of this Ingredient.  # noqa: E501
        :type name: str
        :param allergen_info: The allergen_info of this Ingredient.  # noqa: E501
        :type allergen_info: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'allergen_info': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'allergen_info': 'allergenInfo'
        }
        self._id = id
        self._name = name
        self._allergen_info = allergen_info

    @classmethod
    def from_dict(cls, dikt) -> 'Ingredient':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Ingredient of this Ingredient.  # noqa: E501
        :rtype: Ingredient
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Ingredient.

        Unique identifier for the ingredient.  # noqa: E501

        :return: The id of this Ingredient.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Ingredient.

        Unique identifier for the ingredient.  # noqa: E501

        :param id: The id of this Ingredient.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Ingredient.

        Name of the ingredient.  # noqa: E501

        :return: The name of this Ingredient.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Ingredient.

        Name of the ingredient.  # noqa: E501

        :param name: The name of this Ingredient.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def allergen_info(self) -> str:
        """Gets the allergen_info of this Ingredient.

        Allergen information.  # noqa: E501

        :return: The allergen_info of this Ingredient.
        :rtype: str
        """
        return self._allergen_info

    @allergen_info.setter
    def allergen_info(self, allergen_info: str):
        """Sets the allergen_info of this Ingredient.

        Allergen information.  # noqa: E501

        :param allergen_info: The allergen_info of this Ingredient.
        :type allergen_info: str
        """

        self._allergen_info = allergen_info