# python-flask-server-generated/swagger_server/models/label.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Label(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, description: str=None):  # noqa: E501
        """Label - a model defined in Swagger

        :param id: The id of this Label.  # noqa: E501
        :type id: str
        :param name: The name of this Label.  # noqa: E501
        :type name: str
        :param description: The description of this Label.  # noqa: E501
        :type description: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'description': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description'
        }
        self._id = id
        self._name = name
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'Label':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Label of this Label.  # noqa: E501
        :rtype: Label
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Label.

        Unique identifier for the label.  # noqa: E501

        :return: The id of this Label.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Label.

        Unique identifier for the label.  # noqa: E501

        :param id: The id of this Label.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Label.

        Name of the label.  # noqa: E501

        :return: The name of this Label.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Label.

        Name of the label.  # noqa: E501

        :param name: The name of this Label.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Label.

        Detailed description of the label.  # noqa: E501

        :return: The description of this Label.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Label.

        Detailed description of the label.  # noqa: E501

        :param description: The description of this Label.
        :type description: str
        """

        self._description = description